move i18n to fixtures
