+++ tests/e2e/fixtures/helpers.ts
@@ -3,6 +3,7 @@
 import { Buffer } from 'node:buffer';
 import net from 'node:net';
 import path from 'node:path';
+import { readFileSync } from 'node:fs';
 import { readFile } from 'node:fs/promises';
 import {
   chromium,
@@ -12,6 +13,7 @@ import {
   type Worker,
 } from '@playwright/test';
 import { DIST_DIR, ROOT_DIR } from '../../../esbuild/config';
+import type { TranslationKeys } from '../../../src/shared/helpers';
 
 export type BrowserInfo = { browserName: string; channel: string | undefined };
 export type Background = Worker;
@@ -262,5 +264,62 @@ export async function loadKeysToExtension(
   });
   if (!res || !res.keyId || !res.privateKey || !res.publicKey) {
     throw new Error('Could not load keys to extension');
+  }
+}
+
+type TranslationData = Record<
+  TranslationKeys,
+  { message: string; placeholders?: Record<string, { content: string }> }
+>;
+
+/**
+ * Replacement of browser.i18n.getMessage related APIs
+ */
+export class BrowserIntl {
+  private cache = new Map<string, TranslationData>();
+  private lang = 'en';
+  private pathToExtension: string;
+
+  constructor(browserName: string) {
+    this.pathToExtension = getPathToExtension(browserName);
+  }
+
+  private get(lang: string) {
+    const cached = this.cache.get(lang);
+    if (cached) return cached;
+
+    const filePath = path.join(
+      this.pathToExtension,
+      '_locales',
+      lang,
+      'messages.json',
+    );
+    const data = JSON.parse(readFileSync(filePath, 'utf8')) as TranslationData;
+    this.cache.set(lang, data);
+    return data;
+  }
+
+  getMessage(key: TranslationKeys, substitutions?: string | string[]) {
+    const msg = this.get(this.lang)[key] || this.get('en')[key];
+    if (typeof msg === 'undefined') {
+      throw new Error(`Message not found: ${key}`);
+    }
+
+    let result = msg.message;
+    if (!msg.placeholders) return result;
+
+    if (!substitutions) {
+      throw new Error('Missing substitutions');
+    }
+
+    if (typeof substitutions === 'string') {
+      substitutions = [substitutions];
+    }
+
+    for (const [key, { content }] of Object.entries(msg.placeholders)) {
+      const idx = Number(content.replace('$', ''));
+      result = result.replaceAll(`$${key.toUpperCase()}$`, substitutions[idx]);
+    }
+    return result;
   }
 }
